Extract remaining exported enums
