+++ src/enums/exp_bonus_modifier.ts
@@ -1,0 +1,20 @@
+export enum ExpBonusModifier {
+    POWER_HOUR,
+    BONUS_ARTIST,
+    VOTE,
+    GUESS_STREAK,
+    QUICK_GUESS,
+    MC_GUESS_EASY,
+    MC_GUESS_MEDIUM,
+    MC_GUESS_HARD,
+    ARTIST_GUESS,
+    ARTIST_GUESS_GROUPS_SELECTED,
+    RANDOM_GUESS_BONUS_COMMON,
+    RANDOM_GUESS_BONUS_RARE,
+    RANDOM_GUESS_BONUS_EPIC,
+    RANDOM_GUESS_BONUS_LEGENDARY,
+    BELOW_SONG_COUNT_THRESHOLD,
+    TYPO,
+    HINT_USED,
+    FIRST_GAME_OF_DAY,
+}
