+++ src/enums/locale_type.ts
@@ -1,0 +1,4 @@
+export enum LocaleType {
+    EN = "en",
+    KO = "ko",
+}
