+++ src/helpers/discord_utils.ts
@@ -42,7 +42,8 @@ import TeamScoreboard from "../structures/team_scoreboard";
 import { KmqImages } from "../constants";
 import MessageContext from "../structures/message_context";
 import { REVIEW_LINK, VOTE_LINK } from "../commands/game_commands/vote";
-import { LocaleType, DEFAULT_LOCALE } from "./localization_manager";
+import { LocaleType } from "../enums/locale_type";
+import { DEFAULT_LOCALE } from "./localization_manager";
 import Round from "../structures/round";
 import Session from "../structures/session";
 import MusicSession from "../structures/music_session";
