+++ src/structures/player.ts
@@ -1,7 +1,5 @@
-import {
-    ExpBonusModifier,
-    ExpBonusModifierValues,
-} from "../commands/game_commands/exp";
+import { ExpBonusModifierValues } from "../commands/game_commands/exp";
+import { ExpBonusModifier } from "../enums/exp_bonus_modifier";
 import { getMention } from "../helpers/discord_utils";
 import { bold } from "../helpers/utils";
 import { state } from "../kmq_worker";
